% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_spark.R
\name{dash_spark}
\alias{dash_spark}
\title{Use Shiny and Spark machine learning framework to compare supervised regression models on time series}
\usage{
dash_spark(data = data, x, y, date_column, share_app = FALSE,
  port = NULL)
}
\arguments{
\item{data}{Time serie containing one or more input values and one output value. 
The time serie must be a data.frame or a data.table and must contain at least one time-based column on Date or Posixct format.}

\item{x}{Vector of numerical and categorical input variables used to train and test the model. Each element of x vector must correspond to a data column with either numerical or factor type.}

\item{y}{the numerical output variable to forecast (must correpond to one data column)}

\item{date_column}{the name of time-based column ( must correspond to one data column)}

\item{share_app}{a logical value indicating whether the app must be shared on local LAN}

\item{port}{a four-digit number corresponding to the port the application should listen to. This parameter is necessary only  if share_app option is set to TRUE}
}
\description{
This function creates in one line of code a shareable web app to compare supervised regression model performances.
}
\examples{
\dontrun{
library(dashR)
Sys.setenv(http_proxy="") 
Sys.setenv(http_proxy_user="") 
Sys.setenv(https_proxy_user="")
longley2 <- longley \%>\% mutate(Year = as.Date(as.character(Year),format = "\%Y"))
dash_spark(data =longley2,x = c("GNP_deflator","Unemployed" ,"Armed_Forces","Employed"),
  y = "GNP",date_column = "Year",share_app = TRUE,port = 3952)
}
}
