% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyML_classification.R
\name{shinyML_classification}
\alias{shinyML_classification}
\title{Implement a shiny web app to compare h2o and Spark supervised machine learning models for classification tasks}
\usage{
shinyML_classification(
  data = data,
  y,
  framework = "h2o",
  share_app = FALSE,
  port = NULL
)
}
\arguments{
\item{data}{Time serie containing one or more input values and one output value. 
The time serie must be a data.frame or a data.table and must contain at least one time-based column on Date or POSIXct format.}

\item{y}{the numerical output variable to forecast (must correspond to one data column)}

\item{framework}{the machine learning framework choosed to train and test models (either h2o or Spark). h2o by default.}

\item{share_app}{a logical value indicating whether the app must be shared on local LAN}

\item{port}{a four-digit number corresponding to the port the application should listen to. This parameter is necessary only  if share_app option is set to TRUE}
}
\description{
This function creates in one line of code a shareable web app to compare supervised classification model performances
}
\examples{
\dontrun{
library(shinyML)
shinyML_classification(data = iris,y = "Species",framework = "h2o")
}
}
\author{
Jean Bertin, \email{jean.bertin@gadz.org}
}
